r12490/sofa-dev : FIX: compilation with visual c++ 2005
Former-commit-id: 9913149755f40ad49f8decb629d466ea21eb34a0

+++ framework/sofa/helper/UnitTest.cpp
@@ -58,7 +58,7 @@ bool UnitTest::checkIf( bool testSucceeded, std::string testDescription, unsigne
 
 void UnitTest::initClass()
 {
-    srand( time(NULL) ); // initialize the random number generator using the current time
+    srand( (unsigned int)time(NULL) ); // initialize the random number generator using the current time
 }
